Write header html and link images
I sourced two images that I wanted to use
for the website, got them linked and also
created the list portion that will go in the
top right of the page.
    I know that these list links actually
    need to link to something but I won't
    worry about that now.
Next I put both an "h2" and a "p" for the
other sections.
I listed everything in the header under an
OVERALL class of "header" and then the two
sections as "section-one" and "section-two".

I'm not sure yet if I'll need to remove or reorder those sections once it comes time to write the css.