Write "random info" html section, add images
I wrote an "h3" for the header in this section as I think it was smaller than the header in the previous section. And then I added an ul with linked images and subtext.

I placed all of this under the div "random info" so we'll see if I have to break it up at all.