Practiced and Failed Flexbox design
Over the course of attempting to write my css, I've been able to get things on the landing page in a pseudo-correct order, but not performing exactly how the desired output is designed.

I've been fiddling with changing around the properties, but to no avail. I still am having the problems of the images not sizing correctly. The sections not flexing equally. And the space not being filled. I'm writing this commit as a marker for when I wanted to take a step back, and go over flexbox in more detail before coming back to redo the project.

As of now, I feel as though I'm just going in circles with my adjustments, which I believe is only reenforcing bad habits and understandings.